Fixes to the layout
