+++ assets/css/app.scss
@@ -18,11 +18,9 @@
 @import "./app/components/footer";
 @import "./app/components/hint";
 @import "./app/components/logs_card";
-@import "./app/components/node_selection";
 @import "./app/components/tabular_info";
 @import "./app/components/tabular_page";
 @import "./app/components/progress_section";
-@import "./app/components/refresher";
 @import "./app/components/resource_usage";
 @import "./app/components/modals";
 @import "./app/components/tables";
