--- assets/css/app/components/_node_selection.scss
@@ -1,23 +1,0 @@
-// Node selection dropdown
-#node-selection {
-  position: absolute;
-  top: -4.5rem;
-  right: -0.75rem;
-  padding-right: 0.75rem;
-  padding-top: 0.75rem;
-
-  .custom-select {
-    background-color: change-color($color-gray, $alpha: 0.75);
-    border-radius: $border-radius;
-    border-width: 0;
-    color: black;
-  }
-
-  .input-group-text {
-    border-width: 0;
-    background-color: transparent;
-    color: $color-gray;
-    font-size: 0.9rem;
-    margin-left: -0.5rem;
-  }
-}
