--- assets/css/app/components/_refresher.scss
@@ -1,32 +1,0 @@
-// Refresh interval selection dropdown
-#refresher {
-  position: absolute;
-  bottom: -5.25rem;
-  right: 0;
-  z-index: 2;
-
-  .custom-select {
-    background-color: $color-gray;
-    border-radius: $border-radius;
-    border-width: 0;
-  }
-
-  .input-group-text {
-    border-width: 0;
-    background-color: transparent;
-  }
-}
-
-@media (max-width: map-get($grid-breakpoints, lg)) {
-  #refresher {
-    right: auto;
-    left: auto;
-    text-align: center;
-    width: 100%;
-    bottom: -4.5rem;
-
-    form {
-      display: inline-block;
-    }
-  }
-}
